+++ source/extensions/omni.flux.stage_manager.plugin.filter.usd/docs/CHANGELOG.md
@@ -1,6 +1,17 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.2.0]
+### Added
+- Added `ToggleableUSDFilterPlugin` base filter class
+- Added `LightPrimsFilterPlugin` filter plugin
+
+### Changed
+- Improved `IgnorePrimsFilterPlugin` filtering
+
+### Changed
+- Refactor `OmniPrimsFilterPlugin` to inherit from `ToggleableUSDFilterPlugin`
+
 ## [1.1.1]
 ### Added
 - Added `context_name` field to the USD base