+++ source/extensions/omni.flux.stage_manager.plugin.tree.usd/docs/CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.3.0]
+### Added
+- Added `LightGroupsTreePlugin` plugin
+- Added matching `LightGroupsItem`, `LightGroupsModel`, `LightGroupsDelegate` classes
+
 ## [1.2.0]
 ### Added
 - Added `StageManagerUSDTreeItem`, `StageManagerUSDTreeModel` & `StageManagerUSDTreeDelegate` USD bases